work

+++ src/libs/libwesos-types/include/wesos-types/View.hh
@@ -24,7 +24,7 @@ namespace wesos::types {
   public:
     constexpr View() : m_base(nullptr), m_size(0){};
     constexpr View(Pointer base, usize count) : m_base(base), m_size(count) {}
-    constexpr View(Pointer base, Pointer end) : m_base(base), m_size(end - base) {}
+    constexpr View(Pointer base, Pointer end) : m_base(base), m_size(usize(end - base)) {}
     constexpr View(const View&) = default;
     constexpr View(View&&) = default;
     constexpr auto operator=(const View&) -> View& = default;
